feat: Refactor y optimizar la obtención del estado de sincronización en PhidiasApiService y PhidiasSyncService
