+++ src/services/phidias-api.service.ts
@@ -1,91 +1,6 @@
 // src/services/phidias-api.service.ts
+import { PhidiasPollResponse, PhidiasSyncResult, PhidiasGenericResult, RateLimitConfig, PhidiasConsolidateRecord, PhidiasConsolidateResult } from '@/types/phidias';
 
-export interface PhidiasItem {
-  itemName: string;
-  itemDescription: string;
-  itemvalue: string | number;
-}
-
-export interface PhidiasRecord {
-  id: number;
-  person: number;
-  personFirstname: string;
-  personLastname: string;
-  code: number;
-  author: number;
-  authorFirstname: string;
-  authorLastname: string;
-  last_editor: number;
-  timestamp: number;
-  last_edit: number;
-  items: PhidiasItem[];
-}
-
-export interface PhidiasPollResponse {
-  id: number;
-  year: {
-    id: number;
-    name: string;
-  };
-  name: string;
-  description: string;
-  start_date: number;
-  end_date: number | null;
-  single_record: number;
-  records: PhidiasRecord[];
-}
-
-export interface PhidiasSyncResult {
-  success: boolean;
-  data?: PhidiasPollResponse;
-  error?: string;
-  rateLimited?: boolean;
-  retryAfter?: number;
-}
-
-export interface PhidiasGenericResult {
-  success: boolean;
-  data?: PhidiasPollResponse | PhidiasConsolidateRecord[];
-  error?: string;
-  rateLimited?: boolean;
-  retryAfter?: number;
-}
-
-export interface PhidiasConsolidateRecord {
-  poll_id: number;
-  poll_name: string;
-  person_id: number;
-  person: string;
-  section: string | null;
-  timestamp: string;
-  last_edit: string;
-  author_id: number;
-  author: string;
-  last_editor_id: number;
-  last_editor: string;
-  "Fecha ": string;
-  "Falta según Manual de Convivencia": string;
-  "Descripción de la falta": string;
-  "Acciones Reparadoras": string;
-  "Estudiante con diagnóstico?": string;
-}
-
-export interface PhidiasConsolidateResult {
-  success: boolean;
-  data?: PhidiasConsolidateRecord[];
-  count?: number;
-  error?: string;
-  rateLimited?: boolean;
-  retryAfter?: number;
-}
-
-// Rate limiting configuration
-interface RateLimitConfig {
-  maxRetries: number;
-  baseDelayMs: number;
-  maxDelayMs: number;
-  backoffFactor: number;
-}
 
 class PhidiasApiService {
   private baseUrl: string;
